requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.20.0
prometheus-client==0.16.0
sqlalchemy==1.4.52
databases==0.6.1
httpx==0.24.1
pytest==7.4.0
pytest-asyncio==0.21.0
pydantic==1.10.9
python-dotenv==1.0.0
opentelemetry-api==1.18.0
opentelemetry-sdk==1.18.0
opentelemetry-exporter-jaeger==1.18.0

# All instrumentation MUST use the same version
opentelemetry-instrumentation==0.39b0
opentelemetry-instrumentation-fastapi==0.39b0
opentelemetry-instrumentation-asgi==0.39b0
opentelemetry-instrumentation-requests==0.39b0

opentelemetry-semantic-conventions==0.39b0

requests==2.31.0
